+++ src/state.rs
@@ -55,6 +55,9 @@ pub(crate) enum Command {
 	StylesFolder,
 	SelectTab(usize),
 	CloseTab(usize),
+	Minimize,
+	Maximize,
+	CloseWindow,
 }
 
 #[derive(Default, Clone)]